Improve coding of mvc pattern

+++ src/main/java/com/myproject/movie/models/dtos/responses/PaymentResponseDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 
 @Data
 public class PaymentResponseDto {
-    private Integer id;
-    private Integer bookingId;
+    private Long id;
+    private Long bookingId;
     private Float amount;
     private PaymentMethod paymentMethod;
     private PaymentStatus status;
