+++ src/main/java/com/myproject/movie/repositories/BookingRepository.java
@@ -3,5 +3,5 @@ package com.myproject.movie.repositories;
 import com.myproject.movie.models.entities.Booking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
+public interface BookingRepository extends JpaRepository<Booking, Long> {
 }
