+++ src/main/java/com/myproject/movie/repositories/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.myproject.movie.repositories;
 import com.myproject.movie.models.entities.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {}
+public interface CategoryRepository extends JpaRepository<Category, Long> {}
