+++ src/main/java/com/myproject/movie/repositories/CityRepository.java
@@ -3,5 +3,5 @@ package com.myproject.movie.repositories;
 import com.myproject.movie.models.entities.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Integer> {}
+public interface CityRepository extends JpaRepository<City, Long> {}
