+++ src/main/java/com/myproject/movie/repositories/PaymentRepository.java
@@ -3,5 +3,5 @@ package com.myproject.movie.repositories;
 import com.myproject.movie.models.entities.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
 }
