+++ src/main/java/com/myproject/movie/repositories/ScreeningRepository.java
@@ -3,5 +3,5 @@ package com.myproject.movie.repositories;
 import com.myproject.movie.models.entities.Screening;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScreeningRepository extends JpaRepository<Screening, Integer> {
+public interface ScreeningRepository extends JpaRepository<Screening, Long> {
 }
