+++ src/main/java/com/myproject/movie/repositories/ScreeningSeatRepository.java
@@ -4,6 +4,6 @@ import com.myproject.movie.models.entities.ScreeningSeat;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ScreeningSeatRepository extends JpaRepository<ScreeningSeat, Integer> {
-    List<ScreeningSeat> findByScreeningId(Integer screeningId);
+public interface ScreeningSeatRepository extends JpaRepository<ScreeningSeat, Long> {
+    List<ScreeningSeat> findByScreeningId(Long screeningId);
 }
