+++ src/main/java/com/myproject/movie/services/BookingService.java
@@ -5,5 +5,5 @@ import com.myproject.movie.models.entities.Booking;
 import java.util.List;
 
 public interface BookingService {
-    Booking bookSeats(Integer screeningId, List<Integer> seatIds, Integer userId);
+    Booking bookSeats(Long screeningId, List<Long> seatIds, Long userId);
 }
