+++ src/main/java/com/myproject/movie/utils/CustomPermissionChecker.java
@@ -17,7 +17,7 @@ public class CustomPermissionChecker {
         this.userRepository = userRepository;
     }
 
-    public boolean hasAccess(Authentication authentication, Integer userId) {
+    public boolean hasAccess(Authentication authentication, Long userId) {
         String email = authentication.getName();
         Optional<User> user = userRepository.findByEmail(email);
 
@@ -28,7 +28,7 @@ public class CustomPermissionChecker {
         return checkUserPermission(userEntity, userId);
     }
 
-    private boolean checkUserPermission(User user, Integer userId) {
+    private boolean checkUserPermission(User user, Long userId) {
         return user.getId().equals(userId);
     }
 }
